crypto: address pr comments
