+++ src/crypto/stream.rs
@@ -178,7 +178,6 @@ impl Pull {
 }
 
 #[cfg(test)]
-#[cfg_attr(tarpaulin, skip)]
 mod tests {
     use super::*;
